@@ -56,7 +56,7 @@
     "@types/duo_web_sdk": "2.7.1",
     "@types/firefox-webext-browser": "111.0.1",
     "@types/inquirer": "8.2.6",
-    "@types/jest": "29.5.10",
+    "@types/jest": "29.5.11",
     "@types/jquery": "3.5.29",
     "@types/jsdom": "21.1.6",
     "@types/koa": "2.13.8",
@@ -74,9 +74,9 @@
     "@types/proper-lockfile": "4.1.2",
     "@types/react": "16.14.45",
     "@types/retry": "0.12.2",
-    "@types/zxcvbn": "4.4.1",
-    "@typescript-eslint/eslint-plugin": "6.13.2",
-    "@typescript-eslint/parser": "6.13.2",
+    "@types/zxcvbn": "4.4.4",
+    "@typescript-eslint/eslint-plugin": "6.14.0",
+    "@typescript-eslint/parser": "6.14.0",
     "@webcomponents/custom-elements": "1.6.0",
     "autoprefixer": "10.4.16",
     "base64-loader": "1.0.0",
@@ -87,7 +87,7 @@
     "copy-webpack-plugin": "11.0.0",
     "cross-env": "7.0.3",
     "css-loader": "6.8.1",
-    "electron": "25.9.1",
+    "electron": "27.1.3",
     "electron-builder": "23.6.0",
     "electron-log": "5.0.1",
     "electron-reload": "2.0.0-alpha.1",
@@ -109,7 +109,7 @@
     "gulp-zip": "6.0.0",
     "html-loader": "4.2.0",
     "html-webpack-injector": "1.1.4",
-    "html-webpack-plugin": "5.5.3",
+    "html-webpack-plugin": "5.5.4",
     "husky": "8.0.3",
     "jest-junit": "16.0.0",
     "jest-mock-extended": "3.0.5",
@@ -120,8 +120,8 @@
     "pkg": "5.8.1",
     "postcss": "8.4.31",
     "postcss-loader": "7.3.3",
-    "prettier": "3.1.0",
-    "prettier-plugin-tailwindcss": "0.5.7",
+    "prettier": "3.1.1",
+    "prettier-plugin-tailwindcss": "0.5.9",
     "process": "0.11.10",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -134,14 +134,14 @@
     "style-loader": "3.3.3",
     "tailwindcss": "3.3.5",
     "ts-jest": "29.1.1",
-    "ts-loader": "9.4.4",
+    "ts-loader": "9.5.1",
     "tsconfig-paths-webpack-plugin": "4.1.0",
     "type-fest": "2.19.0",
     "typescript": "4.9.5",
-    "url": "0.11.1",
+    "url": "0.11.3",
     "util": "0.12.5",
     "wait-on": "7.2.0",
-    "webpack": "5.88.2",
+    "webpack": "5.89.0",
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "4.15.1",
     "webpack-node-externals": "3.0.0"
@@ -158,18 +158,18 @@
     "@angular/router": "15.2.10",
     "@koa/multer": "3.0.2",
     "@koa/router": "12.0.0",
-    "@microsoft/signalr": "6.0.21",
-    "@microsoft/signalr-protocol-msgpack": "6.0.21",
+    "@microsoft/signalr": "6.0.25",
+    "@microsoft/signalr-protocol-msgpack": "6.0.25",
     "@ng-select/ng-select": "10.0.4",
     "argon2": "0.31.0",
     "argon2-browser": "1.18.0",
     "big-integer": "1.6.51",
     "bootstrap": "4.6.0",
     "braintree-web-drop-in": "1.40.0",
-    "bufferutil": "4.0.7",
+    "bufferutil": "4.0.8",
     "chalk": "4.1.2",
     "commander": "7.2.0",
-    "core-js": "3.32.0",
+    "core-js": "3.34.0",
     "duo_web_sdk": "github:duosecurity/duo_web_sdk",
     "form-data": "4.0.0",
     "https-proxy-agent": "5.0.1",
