@@ -5,12 +5,14 @@ import { CipherType } from "@bitwarden/common/vault/enums";
 import BrowserPopupUtils from "../../../platform/popup/browser-popup-utils";
 
 import {
+  openViewVaultItemPopout,
   closeAddEditVaultItemPopout,
   closeFido2Popout,
   openAddEditVaultItemPopout,
   openFido2Popout,
   openVaultItemPasswordRepromptPopout,
   VaultPopoutType,
+  closeViewVaultItemPopout,
 } from "./vault-popout-window";
 
 describe("VaultPopoutWindow", () => {
@@ -25,6 +27,34 @@ describe("VaultPopoutWindow", () => {
     jest.clearAllMocks();
   });
 
+  describe("openViewVaultItemPopout", () => {
+    it("opens a popout window that contains a sender tab id query param reference", async () => {
+      const senderTab = { id: 1, windowId: 2 } as chrome.tabs.Tab;
+
+      await openViewVaultItemPopout(senderTab, {
+        cipherId: "cipherId",
+        action: "action",
+      });
+
+      expect(openPopoutSpy).toHaveBeenCalledWith(
+        "popup/index.html#/view-cipher?cipherId=cipherId&senderTabId=1&action=action",
+        {
+          singleActionKey: `${VaultPopoutType.viewVaultItem}_cipherId`,
+          senderWindowId: 2,
+          forceCloseExistingWindows: undefined,
+        },
+      );
+    });
+  });
+
+  describe("closeViewVaultItemPopout", () => {
+    it("closes the view vault item popout window", async () => {
+      await closeViewVaultItemPopout("cipherId");
+
+      expect(closeSingleActionPopoutSpy).toHaveBeenCalledWith(`cipherId`, 0);
+    });
+  });
+
   describe("openVaultItemPasswordRepromptPopout", () => {
     it("opens a popout window that facilitates re-prompting for the password of a vault item", async () => {
       const senderTab = { windowId: 1 } as chrome.tabs.Tab;
@@ -35,7 +65,7 @@ describe("VaultPopoutWindow", () => {
       });
 
       expect(openPopoutSpy).toHaveBeenCalledWith(
-        "popup/index.html#/view-cipher?uilocation=popout&cipherId=cipherId&action=action",
+        "popup/index.html#/view-cipher?cipherId=cipherId&action=action",
         {
           singleActionKey: `${VaultPopoutType.viewVaultItem}_cipherId`,
           senderWindowId: 1,
@@ -52,7 +82,7 @@ describe("VaultPopoutWindow", () => {
       );
 
       expect(openPopoutSpy).toHaveBeenCalledWith(
-        "popup/index.html#/edit-cipher?uilocation=popout&uri=https://jest-testing-website.com",
+        "popup/index.html#/edit-cipher?uri=https://jest-testing-website.com",
         {
           singleActionKey: VaultPopoutType.addEditVaultItem,
           senderWindowId: 1,
@@ -69,7 +99,7 @@ describe("VaultPopoutWindow", () => {
       );
 
       expect(openPopoutSpy).toHaveBeenCalledWith(
-        `popup/index.html#/edit-cipher?uilocation=popout&type=${CipherType.Identity}&uri=https://jest-testing-website.com`,
+        `popup/index.html#/edit-cipher?type=${CipherType.Identity}&uri=https://jest-testing-website.com`,
         {
           singleActionKey: `${VaultPopoutType.addEditVaultItem}_${CipherType.Identity}`,
           senderWindowId: 1,
@@ -86,7 +116,7 @@ describe("VaultPopoutWindow", () => {
       );
 
       expect(openPopoutSpy).toHaveBeenCalledWith(
-        "popup/index.html#/edit-cipher?uilocation=popout&cipherId=cipherId&uri=https://jest-testing-website.com",
+        "popup/index.html#/edit-cipher?cipherId=cipherId&uri=https://jest-testing-website.com",
         {
           singleActionKey: `${VaultPopoutType.addEditVaultItem}_cipherId`,
           senderWindowId: 1,
