@@ -9,6 +9,12 @@ const VaultPopoutType = {
   fido2Popout: "vault_Fido2Popout",
 } as const;
 
+/**
+ * Opens a popout window that facilitates viewing a vault item.
+ *
+ * @param senderTab - The tab that sent the request.
+ * @param cipherOptions - The cipher id and action to perform.
+ */
 async function openViewVaultItemPopout(
   senderTab: chrome.tabs.Tab,
   cipherOptions: {
@@ -18,15 +24,22 @@ async function openViewVaultItemPopout(
   },
 ) {
   const { cipherId, action, forceCloseExistingWindows } = cipherOptions;
-  let promptWindowPath = "popup/index.html#/view-cipher?uilocation=popout";
+  let promptWindowPath = "popup/index.html#/view-cipher";
+  let queryParamToken = "?";
+  const formatQueryString = (key: string, value: string) => {
+    const queryString = `${queryParamToken}${key}=${value}`;
+    queryParamToken = "&";
+    return queryString;
+  };
+
   if (cipherId) {
-    promptWindowPath += `&cipherId=${cipherId}`;
+    promptWindowPath += formatQueryString("cipherId", cipherId);
   }
   if (senderTab.id) {
-    promptWindowPath += `&senderTabId=${senderTab.id}`;
+    promptWindowPath += formatQueryString("senderTabId", String(senderTab.id));
   }
   if (action) {
-    promptWindowPath += `&action=${action}`;
+    promptWindowPath += formatQueryString("action", action);
   }
 
   await BrowserPopupUtils.openPopout(promptWindowPath, {
@@ -36,6 +49,12 @@ async function openViewVaultItemPopout(
   });
 }
 
+/**
+ * Closes the view vault item popout window.
+ *
+ * @param singleActionKey - The single action popout key used to identify the popout.
+ * @param delayClose - The amount of time to wait before closing the popout. Defaults to 0.
+ */
 async function closeViewVaultItemPopout(singleActionKey: string, delayClose = 0) {
   await BrowserPopupUtils.closeSingleActionPopout(singleActionKey, delayClose);
 }
@@ -73,19 +92,25 @@ async function openAddEditVaultItemPopout(
 ) {
   const { cipherId, cipherType } = cipherOptions;
   const { url, windowId } = senderTab;
-
   let singleActionKey = VaultPopoutType.addEditVaultItem;
-  let addEditCipherUrl = "popup/index.html#/edit-cipher?uilocation=popout";
+  let addEditCipherUrl = "popup/index.html#/edit-cipher";
+  let queryParamToken = "?";
+  const formatQueryString = (key: string, value: string) => {
+    const queryString = `${queryParamToken}${key}=${value}`;
+    queryParamToken = "&";
+    return queryString;
+  };
+
   if (cipherId && !cipherType) {
     singleActionKey += `_${cipherId}`;
-    addEditCipherUrl += `&cipherId=${cipherId}`;
+    addEditCipherUrl += formatQueryString("cipherId", cipherId);
   }
   if (cipherType && !cipherId) {
     singleActionKey += `_${cipherType}`;
-    addEditCipherUrl += `&type=${cipherType}`;
+    addEditCipherUrl += formatQueryString("type", String(cipherType));
   }
   if (senderTab.url) {
-    addEditCipherUrl += `&uri=${url}`;
+    addEditCipherUrl += formatQueryString("uri", url);
   }
 
   await BrowserPopupUtils.openPopout(addEditCipherUrl, {
