@@ -2,7 +2,9 @@ import { Injectable } from "@angular/core";
 
 import { ApiService } from "@bitwarden/common/abstractions/api.service";
 import { SelectionReadOnlyRequest } from "@bitwarden/common/admin-console/models/request/selection-read-only.request";
+import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 import { ListResponse } from "@bitwarden/common/models/response/list.response";
+import { ConfigServiceAbstraction } from "@bitwarden/common/platform/abstractions/config/config.service.abstraction";
 
 import { CoreOrganizationModule } from "../../core-organization.module";
 import { GroupView } from "../../views/group.view";
@@ -15,7 +17,27 @@ import { GroupDetailsResponse, GroupResponse } from "./responses/group.response"
   providedIn: "root",
 })
 export class GroupService {
-  constructor(protected apiService: ApiService) {}
+  constructor(
+    protected apiService: ApiService,
+    protected configService: ConfigServiceAbstraction,
+  ) {}
+
+  /**
+   * TODO: This should be replaced with `GroupView.fromResponse` when `FeatureFlag.FlexibleCollections` is removed.
+   **/
+  protected async groupViewFromResponse(response: GroupResponse): Promise<GroupView> {
+    const view = GroupView.fromResponse(response);
+
+    const hasFlexibleCollections = await this.configService.getFeatureFlag(
+      FeatureFlag.FlexibleCollections,
+      false,
+    );
+    if (hasFlexibleCollections) {
+      view.accessAll = false;
+    }
+
+    return view;
+  }
 
   async get(orgId: string, groupId: string): Promise<GroupView> {
     const r = await this.apiService.send(
@@ -26,7 +48,7 @@ export class GroupService {
       true,
     );
 
-    return GroupView.fromResponse(new GroupDetailsResponse(r));
+    return this.groupViewFromResponse(new GroupDetailsResponse(r));
   }
 
   async getAll(orgId: string): Promise<GroupView[]> {
@@ -40,14 +62,17 @@ export class GroupService {
 
     const listResponse = new ListResponse(r, GroupDetailsResponse);
 
-    return listResponse.data?.map((gr) => GroupView.fromResponse(gr)) ?? [];
+    return Promise.all(listResponse.data?.map((gr) => this.groupViewFromResponse(gr))) ?? [];
   }
 }
 
 @Injectable({ providedIn: CoreOrganizationModule })
 export class InternalGroupService extends GroupService {
-  constructor(protected apiService: ApiService) {
-    super(apiService);
+  constructor(
+    protected apiService: ApiService,
+    protected configService: ConfigServiceAbstraction,
+  ) {
+    super(apiService, configService);
   }
 
   async delete(orgId: string, groupId: string): Promise<void> {
@@ -94,7 +119,7 @@ export class InternalGroupService extends GroupService {
       true,
       true,
     );
-    return GroupView.fromResponse(new GroupResponse(r));
+    return this.groupViewFromResponse(new GroupResponse(r));
   }
 
   private async putGroup(
@@ -109,6 +134,6 @@ export class InternalGroupService extends GroupService {
       true,
       true,
     );
-    return GroupView.fromResponse(new GroupResponse(r));
+    return this.groupViewFromResponse(new GroupResponse(r));
   }
 }
