@@ -1,3 +1,4 @@
+import { CommonModule } from "@angular/common";
 import { Component, Injectable, importProvidersFrom } from "@angular/core";
 import { RouterModule } from "@angular/router";
 import {
@@ -27,9 +28,9 @@ import {
   TypographyModule,
   InputModule,
 } from "@bitwarden/components";
-import { PreloadedEnglishI18nModule } from "@bitwarden/web-vault/app/core/tests";
 
-import { HeaderComponent } from "./header.component";
+import { PreloadedEnglishI18nModule } from "../../core/tests";
+import { WebHeaderComponent } from "../header/web-header.component";
 
 @Injectable({
   providedIn: "root",
@@ -66,6 +67,8 @@ class MockProductSwitcher {}
 @Component({
   selector: "dynamic-avatar",
   template: `<bit-avatar [text]="name$ | async"></bit-avatar>`,
+  standalone: true,
+  imports: [CommonModule, AvatarModule],
 })
 class MockDynamicAvatar {
   protected name$ = combineLatest([
@@ -81,7 +84,7 @@ class MockDynamicAvatar {
 
 export default {
   title: "Web/Header",
-  component: HeaderComponent,
+  component: WebHeaderComponent,
   decorators: [
     componentWrapperDecorator(
       (story) => `<div class="tw-min-h-screen tw-flex-1 tw-p-6 tw-text-main">${story}</div>`,
@@ -89,7 +92,6 @@ export default {
     moduleMetadata({
       imports: [
         JslibModule,
-        RouterModule,
         AvatarModule,
         BreadcrumbsModule,
         ButtonModule,
@@ -100,8 +102,9 @@ export default {
         TabsModule,
         TypographyModule,
         NavigationModule,
+        MockDynamicAvatar,
       ],
-      declarations: [HeaderComponent, MockProductSwitcher, MockDynamicAvatar],
+      declarations: [WebHeaderComponent, MockProductSwitcher],
       providers: [
         { provide: StateService, useClass: MockStateService },
         { provide: PlatformUtilsService, useClass: MockPlatformUtilsService },
@@ -126,7 +129,7 @@ export default {
 export const KitchenSink: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="LongTitleeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee" icon="bwi-bug">
+    <app-header title="LongTitleeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee" icon="bwi-bug">
       <bit-breadcrumbs slot="breadcrumbs">
         <bit-breadcrumb>Foo</bit-breadcrumb>
         <bit-breadcrumb>Bar</bit-breadcrumb>
@@ -142,75 +145,75 @@ export const KitchenSink: Story = (args) => ({
         <bit-tab-link route="">Foo</bit-tab-link>
         <bit-tab-link route="#bar">Bar</bit-tab-link>
       </bit-tab-nav-bar>
-    </sm-header>
+    </app-header>
   `,
 });
 
 export const Basic: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug"></sm-header>
+    <app-header title="Foobar" icon="bwi-bug"></app-header>
   `,
 });
 
 export const WithLongTitle: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="LongTitleeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee" icon="bwi-bug"></sm-header>
+    <app-header title="LongTitleeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee" icon="bwi-bug"></app-header>
   `,
 });
 
 export const WithBreadcrumbs: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug" class="tw-text-main">
+    <app-header title="Foobar" icon="bwi-bug" class="tw-text-main">
       <bit-breadcrumbs slot="breadcrumbs">
         <bit-breadcrumb>Foo</bit-breadcrumb>
         <bit-breadcrumb>Bar</bit-breadcrumb>
       </bit-breadcrumbs>
-    </sm-header>
+    </app-header>
   `,
 });
 
 export const WithSearch: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug" class="tw-text-main">
+    <app-header title="Foobar" icon="bwi-bug" class="tw-text-main">
       <input
         bitInput
         placeholder="Ask Jeeves"
         type="text"
       />
-    </sm-header>
+    </app-header>
   `,
 });
 
 export const WithSecondaryContent: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug" class="tw-text-main">
+    <app-header title="Foobar" icon="bwi-bug" class="tw-text-main">
       <button bitButton slot="secondary">Click Me 🎉</button>
-    </sm-header>
+    </app-header>
   `,
 });
 
 export const WithTabs: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug" class="tw-text-main">
+    <app-header title="Foobar" icon="bwi-bug" class="tw-text-main">
       <bit-tab-nav-bar slot="tabs">
         <bit-tab-link route="">Foo</bit-tab-link>
         <bit-tab-link route="#bar">Bar</bit-tab-link>
       </bit-tab-nav-bar>
-    </sm-header>
+    </app-header>
   `,
 });
 
 export const WithCustomTitleComponent: Story = (args) => ({
   props: args,
   template: `
-    <sm-header title="Foobar" icon="bwi-bug" class="tw-text-main">
+    <app-header title="Foobar" icon="bwi-bug" class="tw-text-main">
       <h1 slot="title" class="tw-text-3xl tw-font-semibold" style="font-family: 'Comic Sans MS'">Bitwarden</h1>
-    </sm-header>
+    </app-header>
   `,
 });
