@@ -31,15 +31,15 @@ export class CollectionAdminView extends CollectionView {
     this.assigned = response.assigned;
   }
 
-  override canEdit(org: Organization): boolean {
-    return org?.canEditAnyCollection || (org?.canEditAssignedCollections && this.assigned);
+  override canEdit(org: Organization, flexibleCollectionsEnabled: boolean): boolean {
+    return flexibleCollectionsEnabled
+      ? org?.canEditAnyCollection
+      : org?.canEditAnyCollection || (org?.canEditAssignedCollections && this.assigned);
   }
 
   override canDelete(org: Organization, flexibleCollectionsEnabled: boolean): boolean {
-    if (flexibleCollectionsEnabled) {
-      return org?.canDeleteAnyCollection;
-    } else {
-      return org?.canDeleteAnyCollection || (org?.canDeleteAssignedCollections && this.assigned);
-    }
+    return flexibleCollectionsEnabled
+      ? org?.canDeleteAnyCollection
+      : org?.canDeleteAnyCollection || (org?.canDeleteAssignedCollections && this.assigned);
   }
 }
