@@ -8,15 +8,14 @@ import {
 } from "@bitwarden/components";
 import { CoreOrganizationModule } from "@bitwarden/web-vault/app/admin-console/organizations/core";
 import { DynamicAvatarComponent } from "@bitwarden/web-vault/app/components/dynamic-avatar.component";
+import { HeaderModule } from "@bitwarden/web-vault/app/layouts/header/header.module";
 import { ProductSwitcherModule } from "@bitwarden/web-vault/app/layouts/product-switcher/product-switcher.module";
 import { SharedModule } from "@bitwarden/web-vault/app/shared";
 
 import { AccessPolicySelectorComponent } from "./access-policies/access-policy-selector/access-policy-selector.component";
 import { AccessSelectorComponent } from "./access-policies/access-selector.component";
-import { AccessRemovalDialogComponent } from "./access-policies/dialogs/access-removal-dialog.component";
 import { BulkConfirmationDialogComponent } from "./dialogs/bulk-confirmation-dialog.component";
 import { BulkStatusDialogComponent } from "./dialogs/bulk-status-dialog.component";
-import { HeaderComponent } from "./header.component";
 import { NewMenuComponent } from "./new-menu.component";
 import { OrgSuspendedComponent } from "./org-suspended.component";
 import { ProjectsListComponent } from "./projects-list.component";
@@ -32,17 +31,16 @@ import { SecretsListComponent } from "./secrets-list.component";
     SelectModule,
     DynamicAvatarComponent,
     SearchModule,
+    HeaderModule,
   ],
   exports: [
-    SharedModule,
-    NoItemsModule,
-    AccessRemovalDialogComponent,
-    AccessSelectorComponent,
     AccessPolicySelectorComponent,
-    BulkStatusDialogComponent,
+    AccessSelectorComponent,
     BulkConfirmationDialogComponent,
-    HeaderComponent,
+    BulkStatusDialogComponent,
+    HeaderModule,
     NewMenuComponent,
+    NoItemsModule,
     ProjectsListComponent,
     SearchModule,
     SecretsListComponent,
@@ -50,16 +48,15 @@ import { SecretsListComponent } from "./secrets-list.component";
     SharedModule,
   ],
   declarations: [
-    AccessRemovalDialogComponent,
-    BulkStatusDialogComponent,
+    AccessPolicySelectorComponent,
+    AccessSelectorComponent,
     BulkConfirmationDialogComponent,
-    HeaderComponent,
+    BulkStatusDialogComponent,
+    BulkStatusDialogComponent,
     NewMenuComponent,
+    OrgSuspendedComponent,
     ProjectsListComponent,
     SecretsListComponent,
-    AccessSelectorComponent,
-    OrgSuspendedComponent,
-    AccessPolicySelectorComponent,
   ],
   providers: [],
   bootstrap: [],
