@@ -5,6 +5,7 @@ import {
   InternalOrganizationServiceAbstraction,
   isMember,
 } from "../../abstractions/organization/organization.service.abstraction";
+import { OrganizationUserType } from "../../enums";
 import { OrganizationData } from "../../models/data/organization.data";
 import { Organization } from "../../models/domain/organization";
 
@@ -51,7 +52,7 @@ export class OrganizationService implements InternalOrganizationServiceAbstracti
     return organizations.length > 0;
   }
 
-  async upsert(organization: OrganizationData): Promise<void> {
+  async upsert(organization: OrganizationData, flexibleCollectionsEnabled: boolean): Promise<void> {
     let organizations = await this.stateService.getOrganizations();
     if (organizations == null) {
       organizations = {};
@@ -59,10 +60,10 @@ export class OrganizationService implements InternalOrganizationServiceAbstracti
 
     organizations[organization.id] = organization;
 
-    await this.replace(organizations);
+    await this.replace(organizations, flexibleCollectionsEnabled);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, flexibleCollectionsEnabled: boolean): Promise<void> {
     const organizations = await this.stateService.getOrganizations();
     if (organizations == null) {
       return;
@@ -73,7 +74,7 @@ export class OrganizationService implements InternalOrganizationServiceAbstracti
     }
 
     delete organizations[id];
-    await this.replace(organizations);
+    await this.replace(organizations, flexibleCollectionsEnabled);
   }
 
   get(id: string): Organization {
@@ -102,7 +103,24 @@ export class OrganizationService implements InternalOrganizationServiceAbstracti
     return organizations.find((organization) => organization.identifier === identifier);
   }
 
-  async replace(organizations: { [id: string]: OrganizationData }) {
+  async replace(
+    organizations: { [id: string]: OrganizationData },
+    flexibleCollectionsEnabled: boolean,
+  ) {
+    // If Flexible Collections is enabled, treat Managers as Users and ignore deprecated permissions
+    if (flexibleCollectionsEnabled) {
+      Object.values(organizations).forEach((o) => {
+        if (o.type === OrganizationUserType.Manager) {
+          o.type = OrganizationUserType.User;
+        }
+
+        if (o.permissions != null) {
+          o.permissions.editAssignedCollections = false;
+          o.permissions.deleteAssignedCollections = false;
+        }
+      });
+    }
+
     await this.stateService.setOrganizations(organizations);
     this.updateObservables(organizations);
   }
